+++ workers/unity/Packages/io.improbable.gdk.core/Subscriptions/CallbackManagers/ComponentAddedCallbackManager.cs
@@ -6,7 +6,7 @@ namespace Improbable.Gdk.Subscriptions
 {
     internal class ComponentAddedCallbackManager : ICallbackManager
     {
-        private readonly Callbacks<EntityId> callbacks = new Callbacks<EntityId>();
+        private readonly CallbackCollection<EntityId> callbackCollection = new CallbackCollection<EntityId>();
         private readonly uint componentId;
         private readonly EntityManager entityManager;
         private readonly ComponentUpdateSystem componentUpdateSystem;
@@ -24,19 +24,19 @@ namespace Improbable.Gdk.Subscriptions
             var entities = componentUpdateSystem.GetComponentsAdded(componentId);
             foreach (var entityId in entities)
             {
-                callbacks.InvokeAll(entityId);
+                callbackCollection.InvokeAll(entityId);
             }
         }
 
         public ulong RegisterCallback(Action<EntityId> callback)
         {
-            callbacks.Add(nextCallbackId, callback);
+            callbackCollection.Add(nextCallbackId, callback);
             return nextCallbackId++;
         }
 
         public bool UnregisterCallback(ulong callbackKey)
         {
-            return callbacks.Remove(callbackKey);
+            return callbackCollection.Remove(callbackKey);
         }
     }
 }